use constant dates in update304 tests
As far as I can tell, it isn't supposed to matter what "Last-Modified"
or "Expires" are set to in these tests, so long as we use the same
values across multiple requests. Conceptually, "Last-Modified" ought to
be in the past and "Expires" ought to be in the future, but neither is
supposed to have any effect on the behavior of an HTTP/1.1 cache since
we're setting "Cache-Control: max-age".

I want the tests to be constant data so that exporting them as JSON is
reproducible, so this commit fixes the values which depended on the time
when the tests were generated.

For a time in the past, I picked January 1st of the current year, and I
picked times shortly before signed 32-bit time_t wraps in 2038 to use as
future expiration dates. I don't want this test to fail spuriously if a
cache stores the "Expires" header using 32-bit time_t, although that
could be worth adding a separate test for.