expand templates earlier, when building tests
Currently, if the test author has a typo in the template name, the
template silently expands to nothing. By exporting each template as a
separate JavaScript identifier, any typos in template names are reported
with a stacktrace showing the filename and line number where the error
occurs in the test.

This also marginally simplifies the client without making the tests
harder to read or write. That's useful for anyone trying to duplicate
the client in another language using the output from `export.mjs`.

I tested this by comparing the output from `export.mjs`, both with this
change, and with a copy of the relevant part of `expandTemplates` from
`client.mjs` added to the export module. The keys come out in a
different order but after `python -m json.tool --sort-keys` the two
implementations produce identical output.