don't allow functions in expected_response_headers
Instead of using anonymous functions to check for various more
complicated properties of response headers, this provides a couple of
simple operators that tests can use.

Functions can't be exported as JSON, appearing as `null` in the output
of `JSON.stringify`. So this change is necessary to support exporting
the test suite for use in other languages.

I find the affected test definitions easier to read and understand,
anyway.