update results, except Safari and Fastly
My previous commits changed the formatting of some test failure
messages, so this updates all the results I can do to reflect the new
text.

Also, the previous run had a number of FetchErrors, which I didn't have
a problem with this time, so this cleans up some spurious test failures.

I did rerun the tests against Squid, but the results didn't change.