compare response headers with what server sent
The client wasn't actually checking that the response headers it
received matched the ones the test specified, due to a typo. (See pull
request #62.)

However, checking those headers is complicated because the server uses
information which the client can't know in advance to construct some
headers.

So this patch make the server record what headers it actually sent in
the test state, and then makes the client check the headers it received
against that state.

In addition, sometimes proxies join multiple headers together, or split
a single header into multiple headers. These changes should be allowed,
so this patch normalizes headers into comma-separated single headers
before comparing them.

This isn't ideal because it means the response headers can't be checked
until after all the requests have been issued, so sometimes the error
that's reported is not the one we'd probably prefer to report. But it's
better than the current situation.