permit some response headers to be modified
The general principle I followed is that if a test specifies a response
header which is not syntactically correct or is allowed to be removed by
a proxy, then don't worry if the header didn't make it through intact.

In a few cases, the test is specifically to identify whether a header is
intact after passing through the cache. In those cases, there has to be
an explicit expected_response_headers entry for that header, and it
should not be checked during the response_headers setup phase.

Apache adjusts whitespace in the "Vary" header, which is fine, but I
couldn't come up with a reasonable way to make the tests accept such
modifications, so I just disabled checking that header in the relevant
tests.